+++ grafanalib/core.py
@@ -10,6 +10,7 @@ from attr.validators import instance_of, in_
 import itertools
 import math
 from numbers import Number
+import string
 import warnings
 
 
@@ -1126,6 +1127,36 @@ class Graph(object):
         if self.alert:
             graphObject['alert'] = self.alert
         return graphObject
+
+    def _iter_targets(self):
+        for target in self.targets:
+            yield target
+
+    def _map_targets(self, f):
+        return attr.assoc(self, targets=[f(t) for t in self.targets])
+
+    def auto_ref_ids(self):
+        """Give unique IDs all the panels without IDs.
+
+        Returns a new ``Graph`` that is the same as this one, except all of
+        the metrics have their ``refId`` property set. Any panels which had
+        an ``refId`` property set will keep that property, all others will
+        have auto-generated IDs provided for them.
+        """
+        ref_ids = set([t.refId for t in self._iter_targets() if t.refId])
+        double_candidate_refs = \
+            [p[0] + p[1] for p
+                in itertools.product(string.ascii_uppercase, repeat=2)]
+        candidate_ref_ids = itertools.chain(
+            string.ascii_uppercase,
+            double_candidate_refs,
+        )
+
+        auto_ref_ids = (i for i in candidate_ref_ids if i not in ref_ids)
+
+        def set_refid(t):
+            return t if t.refId else attr.assoc(t, refId=next(auto_ref_ids))
+        return self._map_targets(set_refid)
 
 
 @attr.s